Moved gensym into the builtin package.
