+++ builtins/builtins.go
@@ -12,6 +12,7 @@ package builtins
 import (
 	"fmt"
 	"math"
+	"math/rand"
 	"os"
 	"regexp"
 	"runtime"
@@ -28,6 +29,10 @@ import (
 // These may persist between runs because a regular expression object
 // is essentially constant.
 var regCache map[string]*regexp.Regexp
+
+// symCount is the count of symbols generated by the 'gensym' built-in
+// function.
+var symCount int
 
 // init ensures that our regexp cache is populated
 func init() {
@@ -89,6 +94,10 @@ Example:  (type "string") (type 3) (type type)`})
 	env.Set("help", &primitive.Procedure{F: helpFn, Help: `help returns any help associated with the item specified as the single argument.
 
 \tExample: (print (help print))`})
+
+	env.Set("gensym", &primitive.Procedure{F: gensymFn, Help: `gensym returns a unique symbol, which is guaranteed to be unique.  This is most frequently used inside macro-definitions.
+
+\tExample: (print "%s %s" (gensym) (gensym))`})
 
 	env.Set("keys", &primitive.Procedure{F: keysFn, Help: `keys returns the keys which are present in the specified hash.
 
@@ -384,7 +393,7 @@ func ltFn(env *env.Environment, args []primitive.Primitive) primitive.Primitive 
 }
 
 // equalsFn implements "="
-func equalsFn( env *env.Environment, args []primitive.Primitive) primitive.Primitive {
+func equalsFn(env *env.Environment, args []primitive.Primitive) primitive.Primitive {
 	if len(args) != 2 {
 		return primitive.Error("wrong number of arguments")
 	}
@@ -405,7 +414,7 @@ func equalsFn( env *env.Environment, args []primitive.Primitive) primitive.Primi
 }
 
 // eqFn implements "eq"
-func eqFn( env *env.Environment, args []primitive.Primitive) primitive.Primitive {
+func eqFn(env *env.Environment, args []primitive.Primitive) primitive.Primitive {
 	if len(args) != 2 {
 		return primitive.Error("wrong number of arguments")
 	}
@@ -1028,4 +1037,22 @@ func ordFn(env *env.Environment, args []primitive.Primitive) primitive.Primitive
 		return primitive.Number(float64(rune(s)))
 	}
 	return primitive.Number(0)
+}
+
+// gensymFn is the implementation of (gensym ..)
+func gensymFn(env *env.Environment, args []primitive.Primitive) primitive.Primitive {
+	// symbol characters
+	var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+	// generate prefix
+	b := make([]rune, 5)
+	for i := range b {
+		b[i] = letters[rand.Intn(len(letters))]
+	}
+
+	// generate with count
+	symCount++
+	str := fmt.Sprintf("%s%06d", string(b), symCount)
+	sym := primitive.Symbol(str)
+	return sym
 }
