+++ eval/eval.go
@@ -10,7 +10,6 @@ import (
 	"errors"
 	"flag"
 	"fmt"
-	"math/rand"
 	"regexp"
 	"strconv"
 	"strings"
@@ -37,9 +36,6 @@ type Eval struct {
 
 	// Recurse keeps track of how many times we've recursed
 	recurse int
-
-	// Count of symbols generated by (gensym).
-	symCount int
 }
 
 // New constructs a new evaluator.
@@ -537,24 +533,6 @@ func (ev *Eval) eval(exp primitive.Primitive, e *env.Environment, expandMacro bo
 
 				// Return it.
 				return out
-
-			// (gensym
-			case primitive.Symbol("gensym"):
-
-				// symbol characters
-				var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
-
-				// generate prefix
-				b := make([]rune, 5)
-				for i := range b {
-					b[i] = letters[rand.Intn(len(letters))]
-				}
-
-				// generate with count
-				ev.symCount++
-				str := fmt.Sprintf("%s%06d", string(b), ev.symCount)
-				sym := primitive.Symbol(str)
-				return sym
 
 			// (eval
 			case primitive.Symbol("eval"):
